Add CDI disk preallocation test suite and utility functions
